## Changes
* Related commits (revisions in the serializer) are now prefetched, but not limited to 20 results as before (django does not supports slicing a prefetch, but it is possible in SQL)
  * Avoids 1 SQL request per Push
  * 35516 / 1260041 Push entries are linked to more than 20 commits. 8 are linked to ~10k commits.

## Performance
* Small data (`/api/project/autoland/push/?full=true&count=100&fromchange=<rev_id>`)
  * `master` : 569 ms ± 5.12 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * `e6109c95`: 520 ms ± 4.65 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * Approx 9% time reduction

* Large data (`/api/project/autoland/push/?full=true&count=100`)
  * `master` : 5.69 s ± 63.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * `e6109c95`: 1.87 s ± 69.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
  * Approx 66% time reduction